pyproject.toml
[project]
name = "kindle-to-obsidian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.12.3",
    "pydantic-ai-slim[openai]>=1.3.0",
    "requests>=2.32.5",
    "ruff>=0.14.1",
    "python-dotenv>=1.0.0",
]
